https://public.tableau.com/app/profile/shivani.thakur1849/viz/Netflixtableauproject/Netflix
Developed interactive visualizations in Tableau to identify
patterns and trends. Created a dashboard that effectively communicated the findings and insights to stakeholders. Analyzed user behavior and
preferences, popular genres, and the impact of original content on
viewership. The project aimed to provide actionable insights to help Netflix
improve its content strategy and enhance user experience.